requirements.txt
asgiref==3.8.1
channels==4.2.2
channels_redis==4.2.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
msgpack==1.1.0
redis==5.2.1
sqlparse==0.5.3
